% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDR.to.geometric.mean.R
\name{CDR.to.geometric.mean}
\alias{CDR.to.geometric.mean}
\title{Calculate the Geometric Mean of PRoduction Volume Bins}
\usage{
CDR.to.geometric.mean(x)
}
\arguments{
\item{x}{A string describing the production volume}
}
\value{
A numeric value that is the geometric mean of the bin
}
\description{
Calculate the Geometric Mean of PRoduction Volume Bins
}
\references{
Ring, Caroline L., et al. 
"Consensus modeling of median chemical intake for the US population based on
predictions of exposure pathways." 
Environmental Science & Technology 53.2 (2018): 719-732.
}
\author{
John Wambaugh
}
