% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_seem.R
\name{fast_seem}
\alias{fast_seem}
\title{This is a function for calculate SEEM in parallel}
\usage{
fast_seem(x)
}
\arguments{
\item{x}{A vector of chemical descriptors}
}
\description{
This is a function for calculate SEEM in parallel
}
\references{
Ring, Caroline L., et al. 
"Consensus modeling of median chemical intake for the US population based on
predictions of exposure pathways." 
Environmental Science & Technology 53.2 (2018): 719-732.
}
\author{
John Wambaugh
}
