% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_intakerate.R
\name{lit_intakerate}
\alias{lit_intakerate}
\title{Retrieve intake rates from Ring et al. (2018)}
\usage{
lit_intakerate(chem.ids)
}
\arguments{
\item{chem.ids}{is a vector of DTXSIDs from the CompTox Chemicals Dashboard}
}
\value{
a list with tow entries: matched and unmatched
}
\description{
This function find chemical intake rate predictions that provided as
supplemental information in Ring et al. (2018). These predictions are from 
the SEEM consensus exposure rate model.
}
\details{
Matched is a list of intake rates that were in the supplemental information
Unmatched contaisn all the id's that were not in the supplemental information
of Ring et al. (2018)
}
\references{
Ring, Caroline L., et al. 
"Consensus modeling of median chemical intake for the US population based on predictions of exposure pathways." 
Environmental science & technology 53.2 (2018): 719-732.
}
\author{
John Wambaugh
}
