% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_intakerate.R
\name{predict_intakerate}
\alias{predict_intakerate}
\title{Calculate probababilistithe probabality of exposure via various pathways.}
\usage{
predict_intakerate(
  chem.props,
  file.stub = "new-seem",
  rand.seed = 123456,
  num.draws = 500,
  num.cores = 2,
  model.preds = NULL
)
}
\description{
Calculate probababilistithe probabality of exposure via various pathways.
}
\references{
"Consensus modeling of median chemical intake for the US population based on
predictions of exposure pathways." 
Environmental Science & Technology 53.2 (2018): 719-732.
}
\author{
John Wambaugh and Caroline Ring
}
