% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pathways.R
\name{predict_pathways}
\alias{predict_pathways}
\title{Calculate the probabality of exposure via various pathways.}
\usage{
predict_pathways(chem.props, file.stub = "new-seem", rand.seed = 123456)
}
\arguments{
\item{chem.props}{a N x ?? table of chemical properties to be predicted.}

\item{file.stub}{The text prefix added onto the output files (default "new-seem")}

\item{rand.seed}{The random number generator seed allowing for pseudo-random 
number seuqneces to be replicated.}
}
\value{
SChemical-specific pathway probabilities
}
\description{
This function calculates uses a precomputed random forest model (Ring et al., 2018)
to computer the probabality that chemical exposure occurs via four diferent source-based
pathways: industrial, pesticidal, consumer use, and diet. These pathway probabilities
are used by the SEEM consensus exposure rate model. 

If a chemical is unlikely to have exposure via any of these four pathways, it is considered
to be outside the domain of applicability of the Ring et al. (2018) consensus model.
}
\references{
Ring, Caroline L., et al. 
"Consensus modeling of median chemical intake for the US population based on predictions of exposure pathways." 
Environmental science & technology 53.2 (2018): 719-732.
}
\author{
John Wambaugh and Caroline Ring
}
